% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microb_data.R
\docType{data}
\name{Sample_Data}
\alias{Sample_Data}
\title{Example data, normalized abundance of two taxa and covariates}
\format{
A data frame with 531 rows and 6 variables:
\describe{
  \item{rarefied_taxon1}{rarefied abundance of taxon 1}
  \item{CSS_taxon2}{CSS normalized abundance of taxon 2}
  \item{X}{the clinical variable of interest, binary variable}
  \item{Z1}{covariate 1, continuous variable}
  \item{Z2}{covariate 2, continuous variable}
  \item{Z3}{covariate 3, continuous variable}
}
}
\usage{
Sample_Data
}
\description{
A dataset containing two taxa which have typical abundance profiles highlighting power of ZINQ, including rarefied abundance of taxon 1,
CSS normalized abundance of taxon 2, the clinical variable of interest, and several covariates.
}
\keyword{datasets}
