% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZINQ_combination.R
\name{ZINQ_combination}
\alias{ZINQ_combination}
\title{Combine the marginal p-values}
\usage{
ZINQ_combination(
  input,
  method = "MinP",
  taus = c(0.1, 0.25, 0.5, 0.75, 0.9),
  M = 10000
)
}
\arguments{
\item{input}{An output from \code{\link{ZINQ_tests}}.}

\item{method}{Combination method, "MinP" for MinP test, "Cauchy" for Cauchy combination test; default is "MinP".}

\item{taus}{A grid of quantile levels, must be a subset or equal to that from \code{input}; default is c(0.1, 0.25, 0.5, 0.75, 0.9).}

\item{M}{The number of MC draws from the joint distribution of quantile rank-scores when \code{method} is "MinP"; default is 10000.}
}
\value{
A pvalue, the final p-value of ZINQ
}
\description{
Combine the marginal p-values
}
\details{
\itemize{
  \item Please choose 'MinP' or 'Cauchy' for \code{method}, no other options.
  \item \code{taus} must be a subset or equal to the grid used to produce \code{input}.
}
}
\examples{
library(quantreg)
library(MASS)
n = 300
p <- function(x0, gam0=0.75, gam1=-0.15){
  lc = gam0 + gam1*x0
  exp(lc) / (1 + exp(lc))
}
x = c(rep(0, n), rep(1, n))
w = 0.5 + 1.5*x + (1+0.15*x)*rchisq(2*n,df=1)
b = rbinom(2*n, 1, p(x))
y = w*b
dat = data.frame(y, x)

result = ZINQ_tests(formula.logistic=y~x, formula.quantile=y~x, C="x", data=dat)
ZINQ_combination(result, method="Cauchy")

}
\references{
\itemize{
  \item Ling, W. et al. (2020+). Powerful and robust non-parametric association testing for microbiome data via a zero-inflated quantile approach (ZINQ).
  \item He, Z. et al. (2017). Unified sequence-based association tests allowing for multiple functional annotations and meta-analysis of noncoding variation in metabochip data. The American Journal of HumanGenetics 101(3), 340–352.
  \item Lee, S. et al. (2012). Optimal tests for rare variant effects in sequencing association studies. Biostatistics 13(4), 762–775.
  \item Liu, Y., Xie, J. (2019). Cauchy combination test: a powerful test with analytic p-value calculation under arbitrary dependency structures. Journal of the American Statistical Association, 1–18
}
}
