% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZINQ_tests.R
\name{ZINQ_tests}
\alias{ZINQ_tests}
\title{Marginal tests for the logistic and quantile regression components}
\usage{
ZINQ_tests(
  formula.logistic,
  formula.quantile,
  C,
  y_CorD = "C",
  data,
  taus = c(0.1, 0.25, 0.5, 0.75, 0.9),
  seed = 2020
)
}
\arguments{
\item{formula.logistic}{The full model of logistic regression, e.g., Y ~ X + Y + Z, where Y is zero-inflated.}

\item{formula.quantile}{The full model of quantile regression, can be different from formula.logistic.}

\item{C}{The name(s) of clinical variable(s) of interest, e.g., "Condition" or c("Condition", "Batch").}

\item{y_CorD}{An indicator: use "D" if Y is count, a perturbation from U(0, 1) will be added to the response; use "C" if Y is continuous; default is "C".}

\item{data}{A data.frame: better cleaned and processed, use numeric for Y and binary covariates, use factor for multi-class discrete covariates.}

\item{taus}{A grid of quantile levels, e.g., 0.5 for the median, 0.75 for the 3rd quartile; default is c(0.1, 0.25, 0.5, 0.75, 0.9).}

\item{seed}{A seed for perturbation when \code{y_CorD} is "D"; default is 2020.}
}
\value{
A list
\itemize{
  \item pvalue.logistic - A single p-value from the logistic regression component.
  \item pvalue.quantile - A length(\code{taus}) by 1 vector, a squence of p-values from the quantile regression component.
  \item Sigma.hat - A df x length(\code{taus}) by df x length(\code{taus}) matrix, where df is the dimension of \code{C}, the covariance matrix of quantile rank-scores.
  \item zerorate - The proportion of zeroes in Y.
  \item taus - The grid of quantile levels used.
}
}
\description{
Marginal tests for the logistic and quantile regression components
}
\details{
\itemize{
  \item Compositional data is regarded as continuous, determined by its support.
  \item \code{taus} is a tuning parameter that does not have an efficient selection process yet, try from coarsed to fine grids (e.g., seq(0.1, 0.9, by=0.2) to seq(0.1, 0.9, by=0.1)),
          or try adding more extreme levels (e.g., c(0.25, 0.5, 0.75) to c(0.1, 0.25, 0.5, 0.75, 0.9)), with a goal to keep type I error controlled and boost the power;
        for common taxa, start from the default; for rare taxa, start from c(0.25, 0.5, 0.75).
  \item Quantile rank-score test corrected for zero-inflation is used for the quantile regression component.
  \item If \code{C} is a single continuous or binary covariate, Wald test is used for the logistic regression component, else Rao's score test is used
}
}
\examples{
library(quantreg)
n = 300
p <- function(x0, gam0=0.75, gam1=-0.15){
  lc = gam0 + gam1*x0
  exp(lc) / (1 + exp(lc))
}
x = c(rep(0, n), rep(1, n))
w = 0.5 + 1.5*x + (1+0.15*x)*rchisq(2*n,df=1)
b = rbinom(2*n, 1, p(x))
y = w*b
dat = data.frame(y, x)

ZINQ_tests(formula.logistic=y~x, formula.quantile=y~x, C="x", data=dat)

}
\references{
\itemize{
  \item Ling, W. et al. (2020+). Powerful and robust non-parametric association testing for microbiome data via a zero-inflated quantile approach (ZINQ)
  \item Machado, J.A.F., Silva, J.S. (2005). Quantiles for counts. Journal of the American Statistical Association 100(472), 1226–1237.
}
}
